+++ lib/t/t_mpoption.m
@@ -13,7 +13,7 @@ if nargin < 1
     quiet = 0;
 end
 
-v = 18;
+v = 19;
 
 t_begin(152, quiet);